+++ nrich-excel/src/test/java/net/croz/nrich/excel/service/DefaultExcelReportServiceTest.java
@@ -13,7 +13,7 @@ import org.springframework.util.Assert;
 
 import java.io.File;
 
-import static net.croz.nrich.excel.service.CreateExcelReportRequestGeneratingUtil.createExcelReportRequest;
+import static net.croz.nrich.excel.testutil.ExcelReportRequestGeneratingUtil.createExcelReportRequest;
 import static net.croz.nrich.excel.testutil.PoiDataResolverUtil.createWorkbookAndResolveSheet;
 import static net.croz.nrich.excel.testutil.PoiDataResolverUtil.getRowCellValueList;
 import static org.assertj.core.api.Assertions.assertThat;
